Remove mixins.scss from gitignore
This file used to be compiled, so was ignored. It's not compiled anymore so needs to be unignored.

The file is in version control currently, but gitignore is also used on npm publish, so the file isn't included in the npm bundle, and so compiling the sass of the installed package fails due to a missing file.