Remove styling bound to `#content`
The `#content` container denotes the target of the "skip to main content" link in the govuk template, and so is a functional id. It should not also be used to apply structural style to the page as this means it cannot be moved to other positions to better reflect its functionality.

Instead use the `.main` class which exists on the same container (at present) to apply the structural styling.